Move path_helper to fewshot
